Added PowerShell
Added PowerShell script to help delete randomly, as bat is ill equipped for such things. Merged the settimestamps and new random script in to one script that can be triggered depending on what is passed to it. solved bug where if a screenshot is not found it repeats the count. Thinking also of adding another setting to make the delete.bat silent whilst it is running within the screenshot bat file window.